+++ end_to_end/include/case.hpp
@@ -4,9 +4,11 @@
 
 #pragma once
 
+#include <atomic>
+#include <chrono>
+#include <memory>
 #include <string>
 #include <thread>
-#include <variant>
 #include <vector>
 
 #include "action.hpp"
@@ -21,15 +23,21 @@ class TestCase {
                     const std::vector<std::shared_ptr<TestAction>>& actions)
       : name_(name),
         actions_(actions),
-        kill_reception_(std::make_shared<std::atomic_bool>(false)),
+        ssource_(std::make_shared<std::stop_source>()),
         rx_queue_(std::make_shared<SharedQueue<AstarteMessage>>()) {}
 
   ~TestCase() {
-    if (thread_ && thread_->joinable()) {
-      *kill_reception_ = true;
-      thread_->join();
+    if (ssource_) {
+      ssource_->request_stop();
     }
   }
+
+  // TestaCase is not copyable
+  TestCase(const TestCase&) = delete;
+  auto operator=(const TestCase&) -> TestCase& = delete;
+  // TestCase is movable
+  TestCase(TestCase&&) = default;
+  auto operator=(TestCase&& other) -> TestCase& = default;
 
   void configure_curl(const std::string& appengine_url, const std::string& appengine_token,
                       const std::string& realm, const std::string& device_id) {
@@ -37,17 +45,21 @@ class TestCase {
       action->configure_curl(appengine_url, appengine_token, realm, device_id);
     }
   }
+
   void attach_device(const std::shared_ptr<AstarteDeviceGRPC>& device) {
     for (const auto& action : actions_) {
-      action->attach_device(device, rx_queue_, kill_reception_);
+      action->attach_device(device, rx_queue_, ssource_);
     }
     device_ = device;
   }
+
   void start() {
     if (!thread_) {
-      thread_ = std::make_shared<std::thread>(&TestCase::reception_handler, this);
+      thread_ =
+          std::make_shared<std::jthread>(&TestCase::reception_handler, this, ssource_->get_token());
     }
   }
+
   void execute() {
     spdlog::info("Executing test case: {}.", name_);
     for (const auto& action : actions_) {
@@ -59,12 +71,13 @@ class TestCase {
   std::vector<std::shared_ptr<TestAction>> actions_;
   std::string name_;
   std::shared_ptr<AstarteDeviceGRPC> device_;
-  std::shared_ptr<std::atomic_bool> kill_reception_;
+  // master stop source to cleanly join all testcases threads
+  std::shared_ptr<std::stop_source> ssource_;
   std::shared_ptr<SharedQueue<AstarteMessage>> rx_queue_;
-  std::shared_ptr<std::thread> thread_;
+  std::shared_ptr<std::jthread> thread_;
 
-  void reception_handler() {
-    while (!(*kill_reception_)) {
+  void reception_handler(std::stop_token token) {
+    while (!token.stop_requested()) {
       auto incoming = device_->poll_incoming(std::chrono::milliseconds(100));
       if (incoming.has_value()) {
         AstarteMessage msg(incoming.value());